requirements.txt
colorama>=0.4.6
google-genai>=0.1.0
python-dotenv>=1.0.0
